pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
requests = "^2.25.1"
pymongo = {extras = ["srv"], version = "^4.0.1"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
selenium = "^3.141.0"
pylint = "^2.12.2"
autopep8 = "^1.6.0"
mongomock = "^4.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
